fixed: link in delete task notification; refactoring code - deadline notification

+++ telebot/models.py
@@ -5,7 +5,7 @@ from pulsewave import settings
 
 class TeleBotID(models.Model):
     user = models.OneToOneField(to=settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-    telegram_id = models.PositiveIntegerField()
+    telegram_id = models.PositiveBigIntegerField()
     name = models.CharField(max_length=32)
 
     def __repr__(self):
